support multiprocessor for extract aicity data

+++ tools/dataset_converters/aicity/vid2imgs.py
@@ -2,7 +2,9 @@ import os, shutil
 from warnings import warn
 from tqdm import tqdm
 from argparse import ArgumentParser
-
+import multiprocessing as mp
+from concurrent.futures import ThreadPoolExecutor
+from math import ceil
 import mmcv
 
 
@@ -10,6 +12,7 @@ def parse_args():
     parser = ArgumentParser()
     parser.add_argument("data_dir", help="Path to the data directory")
     parser.add_argument("--fps", type=int, default=5, help="FPS to extract images at. Default: 5")
+    parser.add_argument("--num_workers", type=int, default=mp.cpu_count(), help="Number of workers to use. Default: all available cpus")
 
     return parser.parse_args()
 
@@ -25,14 +28,21 @@ def extract_images(vid_path: str, imgs_dir: str, fps: int):
     if not os.path.exists(vid_path):
         print(f"Missing video: {vid_path}")
 
+    reader = mmcv.VideoReader(vid_path)
+    frame_step = int(reader.fps / fps)
+
     if os.path.exists(imgs_dir):
-        warn(f"Images already extracted: {imgs_dir}. Removing...")
-        shutil.rmtree(imgs_dir)
+        extracted_imgs = os.listdir(imgs_dir)
+        num_frames = ceil(reader.frame_cnt / frame_step)
+
+        if len(extracted_imgs) == num_frames:
+            print(f"Images already extracted: {imgs_dir}. Skipping...")
+            return
+        elif len(extracted_imgs) > 0:
+            print(f"Mismatch in number of extracted images. Number of extracted images: {len(extracted_imgs)}. Expected {num_frames}")
+            shutil.rmtree(imgs_dir)
     os.makedirs(imgs_dir)
 
-    reader = mmcv.VideoReader(vid_path)
-    
-    frame_step = int(reader.fps / fps)
 
     for frame_id in range(0, reader.frame_cnt, frame_step):
         frame = reader.get_frame(frame_id)
@@ -41,19 +51,21 @@ def extract_images(vid_path: str, imgs_dir: str, fps: int):
 def main():
     args = parse_args()
 
-    for subset in ("train", "validation", "test"):
+    def extract_images_for_camera(camera_dir):
+        if camera_dir.is_dir():
+            vid_path = os.path.join(camera_dir.path, "video.mp4")
+            imgs_dir = os.path.join(camera_dir.path, "imgs")
+
+            extract_images(vid_path, imgs_dir, args.fps)
+
+    for subset in ("train", "validation"):
         subset_dir = os.path.join(args.data_dir, subset)
 
-        print(f'Extracing images from {subset} set')
+        print(f'Extracting images from {subset} set')
         for scene_dir in tqdm(os.scandir(subset_dir)):
             if scene_dir.is_dir():
-                for camera_dir in os.scandir(scene_dir):
-                    if camera_dir.is_dir():
-                        vid_path = os.path.join(camera_dir.path, "video.mp4")
-                        imgs_dir = os.path.join(camera_dir.path, "imgs")
-
-                        extract_images(vid_path, imgs_dir, args.fps)
-
+                with ThreadPoolExecutor(max_workers=args.num_workers) as executor:
+                    executor.map(extract_images_for_camera, os.scandir(scene_dir.path))
 
 if __name__ == "__main__":
     main()
